Manage closures as intervals, Fix dates inside calendar detail, add operators permission checks, fix availabilities API

+++ src/AppBundle/Controller/CalendarsController.php
@@ -6,6 +6,7 @@ use AppBundle\Entity\Calendar;
 use AppBundle\Entity\Meeting;
 use Doctrine\DBAL\Exception\ForeignKeyConstraintViolationException;
 use Doctrine\ORM\QueryBuilder;
+use Omines\DataTablesBundle\Adapter\ArrayAdapter;
 use Omines\DataTablesBundle\Adapter\Doctrine\ORMAdapter;
 use Omines\DataTablesBundle\Column\DateTimeColumn;
 use Omines\DataTablesBundle\Column\NumberColumn;
@@ -34,16 +35,35 @@ class CalendarsController extends Controller
   public function indexCalendarsAction(Request $request)
   {
     $em = $this->getDoctrine()->getManager();
-    $items = $em->getRepository('AppBundle:Calendar')->findAll();
+    $data = [];
+
+    $builder = $em->createQueryBuilder();
+    $query = $builder
+      ->select('calendar', 'moderators')
+      ->from(Calendar::class, 'calendar')
+      ->leftJoin('calendar.moderators', 'moderators')
+      ->where('calendar.owner = :owner')
+      ->Orwhere('moderators.id = :operatore')
+      ->setParameter('owner', $this->getUser())
+      ->setParameter('operatore', $this->getUser())
+      ->getQuery();
+
+    foreach ($query->getResult() as $calendarEntry) {
+      $data[] = array(
+        'title' => $calendarEntry->getTitle(),
+        'id' => $calendarEntry->getId(),
+        'owner'=>$calendarEntry->getOwner()->getUsername(),
+        'isModerated'=>$calendarEntry->getIsModerated(),
+      );
+    }
 
     $table = $this->createDataTable()
       ->add('title', TextColumn::class, ['label' => 'Titolo', 'propertyPath' => 'Titolo', 'render' => function ($value, $calendar) {
         return sprintf('<a href="%s">%s</a>', $this->generateUrl('operatori_calendar_show', [
-          'calendar' => $calendar->getId()
-        ]), $calendar->getTitle());
+          'calendar' => $calendar['id']
+        ]), $calendar['title']);
       }])
-      //->add('title', TextColumn::class, ['label'=>'Titolo', 'searchable'=>true])
-      ->add('owner', TextColumn::class, ['label' => 'Proprietario', 'field' => 'owner.username', 'searchable' => true])
+      ->add('owner', TextColumn::class, ['label' => 'Proprietario', 'searchable' => true])
       ->add('isModerated', TextColumn::class, ['label' => 'Moderazione', 'render' => function ($value, $calendar) {
         if ($value) {
           return sprintf('<span class="badge badge-outline-primary"><svg class="icon icon-sm icon-primary"><use xlink:href="/bootstrap-italia/dist/svg/sprite.svg#it-locked"></use></svg>Richiede moderazione</span>');
@@ -63,16 +83,13 @@ class CalendarsController extends Controller
           $this->generateUrl('operatori_calendar_delete', ['id' => $value])
         );
       }])
-      ->createAdapter(ORMAdapter::class, [
-        'entity' => Calendar::class
-      ])
+      ->createAdapter(ArrayAdapter::class, $data)
       ->handleRequest($request);
 
     if ($table->isCallback()) {
       return $table->getResponse();
     }
     return array(
-      'items' => $items,
       'datatable' => $table
     );
   }
@@ -128,8 +145,12 @@ class CalendarsController extends Controller
    */
   public function deleteCalendarAction(Request $request, Calendar $calendar)
   {
-    try {
+    if ($calendar->getOwner() != $this->getUser() && $calendar->getIsModerated() && !$calendar->getModerators()->contains($this->getUser())) {
+      $this->addFlash('error', 'Non possiedi i permessi per eliminare questo calendario');
+      return $this->redirectToRoute('operatori_calendars_index');
+    }
 
+    try {
       $em = $this->getDoctrine()->getManager();
       $em->remove($calendar);
       $em->flush();
@@ -169,6 +190,11 @@ class CalendarsController extends Controller
    */
   public function editCalendarAction(Request $request, Calendar $calendar)
   {
+    if ($calendar->getOwner() != $this->getUser() && $calendar->getIsModerated() && !$calendar->getModerators()->contains($this->getUser())) {
+      $this->addFlash('error', 'Non possiedi i permessi per modificare questo calendario');
+      return $this->redirectToRoute('operatori_calendars_index');
+    }
+
     $form = $this->createForm('AppBundle\Form\CalendarBackofficeType', $calendar);
     $form->handleRequest($request);
 
@@ -196,6 +222,10 @@ class CalendarsController extends Controller
         foreach ($toDelete as $deleteId => $openingHour) {
           $em->remove($openingHour);
         }
+        if (!$calendar->getIsModerated()) {
+          // remove moderators
+          $calendar->setModerators([]);
+        }
         $em->persist($calendar);
         $em->flush();
 
@@ -219,6 +249,11 @@ class CalendarsController extends Controller
    */
   public function showCalendarAction(Request $request, Calendar $calendar)
   {
+    if ($calendar->getOwner() != $this->getUser() && $calendar->getIsModerated() && !$calendar->getModerators()->contains($this->getUser())) {
+      $this->addFlash('error', 'Non possiedi i permessi per visualizzare questo calendario');
+      return $this->redirectToRoute('operatori_calendars_index');
+    }
+
     $statuses = [
       Meeting::STATUS_PENDING => 'In attesa di approvazione',
       Meeting::STATUS_APPROVED => 'Approvato',
@@ -228,18 +263,18 @@ class CalendarsController extends Controller
     ];
 
     $table = $this->createDataTable()
-      ->add('fromTime', DateTimeColumn::class, ['label'=>'Data', 'format'=> 'Y-m-d'])
-      ->add('id', TextColumn::class, ['label'=>'Orario', 'render'=> function($value, $meeting) {
+      ->add('fromTime', DateTimeColumn::class, ['label' => 'Data', 'format' => 'Y-m-d'])
+      ->add('id', TextColumn::class, ['label' => 'Orario', 'render' => function ($value, $meeting) {
         return sprintf('%s - %s', $meeting->getFromTime()->format('H:i'), $meeting->getToTime()->format('H:i'));
       }])
-      ->add('Nome e Cognome', TextColumn::class, ['label' => 'Utente', 'searchable'=>true, 'render' => function ($value, $meeting) {
+      ->add('Nome e Cognome', TextColumn::class, ['label' => 'Utente', 'searchable' => true, 'render' => function ($value, $meeting) {
         if ($meeting->getUser()) return $meeting->getUser()->getUsername();
         else return 'Utente Anonimo';
       }])
-      ->add('email', TextColumn::class, ['label' => 'Email', 'searchable'=>true, 'render' => function ($value, $meeting) {
+      ->add('email', TextColumn::class, ['label' => 'Email', 'searchable' => true, 'render' => function ($value, $meeting) {
         return $value ? sprintf('<a href="mailto:%s"><div class="text-truncate">%s</div></a>', $value, $value) : '---';
       }])
-      ->add('phoneNumber', TextColumn::class, ['label' => 'Recapito', 'render' => function($value, $meeting) {
+      ->add('phoneNumber', TextColumn::class, ['label' => 'Recapito', 'render' => function ($value, $meeting) {
         return $value ? $value : '---';
       }])
       ->add('rescheduled', NumberColumn::class, ['label' => 'Rinvii'])
@@ -255,8 +290,7 @@ class CalendarsController extends Controller
             ->select('meeting')
             ->from(Meeting::class, 'meeting')
             ->where('meeting.calendar = :calendar')
-            ->setParameter('calendar', $calendar)
-          ;
+            ->setParameter('calendar', $calendar);
         },
       ])
       ->handleRequest($request);
@@ -319,7 +353,7 @@ class CalendarsController extends Controller
       }
       $events[] = [
         'id' => $meeting->getId(),
-        'title' => $meeting->getName(),
+        'title' => $meeting->getName() ? $meeting->getName() : 'Nome non fornito',
         'name' => $meeting->getName(),
         'start' => $meeting->getFromTime()->format('c'),
         'end' => $meeting->getToTime()->format('c'),
@@ -351,11 +385,18 @@ class CalendarsController extends Controller
       $events = array_merge($events, $openingHour->getInterval());
       $minDuration = min($minDuration, $openingHour->getMeetingMinutes());
     }
+
+    // Check permissions if calendar is moderated
+    if (!$calendar->getIsModerated() || ($this->getUser() == $calendar->getOwner() || $calendar->getModerators()->contains($this->getUser())))
+      $canEdit = true;
+    else $canEdit = false;
+
     return array(
       'calendar' => $calendar,
+      'canEdit' => $canEdit,
       'delete_form' => $deleteForm->createView(),
       'events' => array_values($events),
-      'statuses'=> $statuses,
+      'statuses' => $statuses,
       'minDuration' => $minDuration,
       'datatable' => $table
     );
